modify: request에 관광지 정보 포함
